Notes from following the instructions
moniker is auto generated
missing "" on multiple bootnodes (maybe not required)

This doesnt seem to be the latest changes as the section "To check if Heimdall is synced" is not reflected here. It would be nice to say heimdall sync could take approx 4 hours or so. Check `latest_block_height` and https://mumbai-explorer.matic.today/blocks to get an idea of how far it is from syncing. (assuming `latest_block_height`is the latest block)